requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.16.1
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
gunicorn==25.1.0
packaging==26.0
